requirements.txt
streamlit==1.35.0
mcp==1.12.0
apscheduler==3.10.4
python-dotenv==1.0.0
openai==0.28.1